+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - [TD-5038] Links between structures
+- [TD-5037] Added diff information for draft and pending_aproval structure notes
+  in API response
 
 ### Fixed